# frontend: use zero-allocation response decoder in shard active series middleware 