--- a/pkg/frontend/querymiddleware/shard_active_series.go
+++ b/pkg/frontend/querymiddleware/shard_active_series.go
@@ -538,7 +538,7 @@ func (s *shardActiveSeriesMiddleware) writeMergedResponseWithZeroAllocationDecod
 		}
 
 		// Reuse stream buffer.
-		reuseActiveSeriesDataChunkBuffer(streamBuf)
+		reuseActiveSeriesDataStreamBuffer(streamBuf)
 	}
 	stream.WriteArrayEnd()