… CHANGELOG.md
… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… pkg/frontend/querymiddleware/shard_active_series_response_decoder_test.go
… pkg/frontend/querymiddleware/shard_active_series_test.go