<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

At certain times we've observed in ops huge spikes of allocations during the deserialization/processing of each of the query responses when merging the results in the shard active series frontend middleware, occasionally reaching values close to `~80%` of total allocations.

![Screenshot 2024-03-21 at 09 28 42](https://github.com/grafana/mimir/assets/888899/b2338344-817d-4cfa-a8aa-786c9489c430)

While this condition does not critically compromise the service, nor does imply a customer impact, such spikes translate into high CPU consumption that can negatively impact the TCO.

This PR aims to solve this problem through the implementation of a zero-allocation decoder, specifically adapted to the use case of said middleware, that borrows certain functions from the json-iter library, currently in use.

Broadly speaking, the way this new decoder works is by copying all contents of the response `data` key into an internal buffer that is subsequently reused once is received and consumed by the merger goroutine, rather than copying each of the strings associated with all label-value pairs.

_Below are the comparative benchmark results before and after the change._

```
name                                                                         old time/op    new time/op    delta
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-4-8           8.93µs ± 0%    7.51µs ± 3%   -15.94%  (p=0.016 n=4+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-16-8          15.9µs ± 1%    15.9µs ± 2%      ~     (p=0.690 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-64-8          45.0µs ± 1%    43.3µs ± 2%    -3.75%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-128-8         77.8µs ± 0%    79.2µs ± 3%      ~     (p=0.222 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-4-8         14.5µs ± 1%    12.7µs ± 1%   -12.09%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-16-8        25.4µs ± 3%    24.0µs ± 1%    -5.66%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-64-8        59.9µs ± 1%    61.2µs ± 0%    +2.22%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-128-8       91.7µs ± 2%    98.7µs ± 2%    +7.58%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-4-8       1.90ms ± 0%    0.51ms ± 0%   -72.95%  (p=0.016 n=5+4)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-16-8      7.22ms ± 0%    2.10ms ± 8%   -70.95%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-64-8      28.3ms ± 1%     7.7ms ±11%   -72.63%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-128-8     56.1ms ± 1%    16.0ms ±12%   -71.52%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-4-8      18.8ms ± 0%     5.0ms ± 0%   -73.26%  (p=0.016 n=5+4)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-16-8     72.2ms ± 0%    22.6ms ±11%   -68.65%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-64-8      281ms ± 1%      81ms ± 4%   -71.23%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-128-8     558ms ± 3%     150ms ± 1%   -73.13%  (p=0.008 n=5+5)

name                                                                         old alloc/op   new alloc/op   delta
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-4-8           1.98kB ± 0%    1.80kB ± 2%    -9.07%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-16-8          3.90kB ± 0%    2.91kB ± 1%   -25.31%  (p=0.016 n=4+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-64-8          11.6kB ± 0%     8.7kB ± 1%   -24.79%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-128-8         22.0kB ± 0%    19.8kB ± 1%    -9.67%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-4-8         2.75kB ± 5%    2.55kB ± 5%    -7.41%  (p=0.016 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-16-8        4.91kB ± 6%    3.92kB ± 6%   -20.26%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-64-8        13.0kB ± 4%    11.2kB ± 4%   -14.07%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-128-8       23.5kB ± 4%    24.3kB ±10%      ~     (p=0.222 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-4-8        406kB ± 0%      95kB ± 0%   -76.58%  (p=0.016 n=5+4)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-16-8      1.37MB ± 0%    0.12MB ± 2%   -91.19%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-64-8      5.25MB ± 0%    0.43MB ±24%   -91.80%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-128-8     10.5MB ± 0%     1.2MB ±59%   -88.51%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-4-8      3.29MB ± 0%    0.88MB ± 0%   -73.38%  (p=0.016 n=5+4)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-16-8     12.9MB ± 0%     2.0MB ± 3%   -84.61%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-64-8     51.7MB ± 0%     5.4MB ±12%   -89.61%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-128-8     104MB ± 0%      17MB ±14%   -83.32%  (p=0.008 n=5+5)

name                                                                         old allocs/op  new allocs/op  delta
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-4-8             60.0 ± 0%      25.0 ± 0%   -58.33%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-16-8             192 ± 0%        49 ± 0%   -74.48%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-64-8             720 ± 0%       145 ± 0%   -79.86%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=none/num-responses-128-8          1.42k ± 0%     0.27k ± 0%   -80.83%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-4-8           67.0 ± 0%      32.0 ± 0%   -52.24%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-16-8           199 ± 0%        56 ± 0%   -71.86%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-64-8           727 ± 0%       152 ± 0%   -79.09%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/encoding=snappy/num-responses-128-8        1.43k ± 0%     0.28k ± 0%   -80.43%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-4-8        32.1k ± 0%      0.0k ± 0%   -99.91%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-16-8        128k ± 0%        0k ± 0%      ~     (p=0.079 n=4+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-64-8        513k ± 0%        0k ± 0%   -99.97%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=1_000/num-responses-128-8      1.03M ± 0%     0.00M ± 1%   -99.97%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-4-8        320k ± 0%        0k ± 0%      ~     (p=0.079 n=4+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-16-8      1.28M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=5+4)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-64-8      5.12M ± 0%     0.00M ± 2%  -100.00%  (p=0.008 n=5+5)
ActiveSeriesMiddlewareMergeResponses/seriesCount=10_000/num-responses-128-8     10.2M ± 0%      0.0M ± 4%  -100.00%  (p=0.008 n=5+5)
```

_Raw benchmark results_:
* [old.bench](https://gist.github.com/ortuman/abfb50aace899ca08bf437cd0689ad12)  
* [new.bench](https://gist.github.com/ortuman/912707436a5c828d95cb493e4285d952)

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/mimir-squad/issues/1870

#### Checklist

- [X] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.