@flxbk 2024-03-21
Cool, thanks for working on this! I'll try to give it a closer look soon, but one thing that occurred to me: Do you think it's worth and feasible putting this behind a feature flag to make rollback easier in case any unexpected issues appear? Not sure how risky you deem this change.