@ortuman 2024-03-21
> Do you think it's worth and feasible putting this behind a feature flag to make rollback easier in case any unexpected issues appear?

I believe it makes complete sense and is perfectly viable. Once we've tested that everything works as expected we can get rid of the old code. 👍 

I'll see to rework the PR to allow both code paths. In the meantime, I'll mark the PR as a draft again.