@ortuman 2024-03-21
Added new `-query-frontend.use-shard-active-series-zero-allocation-decoder` experimental option to enable/disable this feature at will. 

_@flxbk @Logiraptor this is ready for review now_