@ortuman 2024-04-01
> I do worry about debugging this if it ever breaks, so I added a fuzz test which compares this PR to the stdlib json result. The first failure is the input "{\"data\":[]" due to the missing }. The code in the PR accepts that invalid json. I understand that this code is extremely fast, but it makes me nervous at the same time. Do you think making the code fail on invalid json will eliminate the performance gains?

If I remember correctly, the current implementation would also accept the input `{"data": [...]`, as the processing of the querier response stops as soon as the content of the `data` key has been processed.

Also, after the change in 23e736b4a289ba1efaa27651d984402c3e6e1fd2, at least now we ensure that the content of said key at least takes the form `[{...},{...}, ...]`, although it is true that the objects contained in the array can contain any random data, and such content would be propagated to the frontend. Which, on the other hand, I don't see as a cause for concern. If the querier returns corrupt data, it would make sense for the frontend to propagate such data, thereby highlighting the problem instead of covering it up.